+++ PathObjectivesAndConstraints/src/CubicEquationSolver.hpp
@@ -6,15 +6,15 @@
 
 namespace CubicEquationSolver
 {
-    std::array<double,3> solve_equation(double &a_term, double &b_term, double &c_term, double &d_term);
+    std::array<float,3> solve_equation(float &a_term, float &b_term, float &c_term, float &d_term);
 
-    std::array<double,3> solve_cubic_equation(double &a_term, double &b_term, double &c_term, double &d_term);
+    std::array<float,3> solve_cubic_equation(float &a_term, float &b_term, float &c_term, float &d_term);
 
-    std::array<double,2> solve_quadratic_equation(double &b_term, double &c_term, double &d_term);
+    std::array<float,2> solve_quadratic_equation(float &b_term, float &c_term, float &d_term);
 
-    double solve_linear_equation(double &c_term, double &d_term);
+    float solve_linear_equation(float &c_term, float &d_term);
 
-    static const double DOUBLEMAX = std::numeric_limits<double>::max();
+    static const float DOUBLEMAX = std::numeric_limits<float>::max();
 }
 
 #endif
